+++ src/ws/mod.rs
@@ -1,3 +1,3 @@
 pub mod terminal;
 
-pub use terminal::ws_terminal_handler;
+pub use terminal::{ws_external_terminal_handler, ws_terminal_handler};